Safari-611.1.10/LayoutTests/editing/pasteboard/datatransfer-items-paste-plaintext-expected.txt

This tests dataTransfer.items. To manually test, click on "Copy text" and paste (Command+V on Mac Control+V elsewhere).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS clipboardData.items.length is 1
PASS clipboardData.items[0].kind is "string"
PASS clipboardData.items[0].type is "text/plain"
PASS clipboardData.items[0].getAsFile() is null
PASS clipboardData.items[0].getAsString(checkContent) is undefined
PASS "hello, world" is "hello, world"
PASS successfullyParsed is true

TEST COMPLETE
